+++ view/stache/stache_test.js
@@ -4130,4 +4130,22 @@ steal("can-simple-dom", "can/util/vdom/build_fragment","can/view/stache", "can/v
 	}
 
 
+	test('Custom attribute callbacks are called when in a conditional within a live section', 8, function () {
+		can.view.attr('test-attr', function(el, attrData) {
+			ok(true, "test-attr called");
+			equal(attrData.attributeName, 'test-attr', "attributeName set correctly");
+			ok(attrData.scope, "scope isn't undefined");
+			ok(attrData.options, "options isn't undefined");
+		});
+
+		var state = new can.Map({
+			showAttr: true
+		});
+
+		var template = can.stache('<button id="find-me" {{#if showAttr}}test-attr{{/if}}></button>');
+		template(state);
+
+		state.attr('showAttr', false);
+		state.attr('showAttr', true);
+	});
 });